chore(widget-chat): rename ChatWidget to chat-widget
