+++ packages/widget-chat/src/chat-widget.js
